{"types":"C","runs":47}

+++ problem15.java
@@ -5,7 +5,7 @@ public class problem15 extends ConsoleProgram {
 	public void run() {
 		int firstNum = readInt("First number is: ");
 		int secondNum = readInt("Second number is: ");
-		double average = (firstNum + secondNum) / 2;
+		float average = (firstNum + secondNum) / 2;
 		println("Arithmetic average is: " + average);		
 	}
 }
